requirements-dev.txt
# Development and testing dependencies
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-env>=1.0.0
pytest-mock>=3.11.1
httpx>=0.24.0  # For async test client support
python-dotenv
